+++ sagemaker-mlops/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v1.4.0 (2026-01-22)
+
+* feat: add emr-serverless step for SageMaker Pipelines
+
 ## v1.3.1 (2026-01-12)
 
 * sagemaker-mlops bug fix - Correct source code 'dependencies' parameter to 'requirements'